pine: Add wifi mac generator
The default qcacld-3.0 implementation looks for the MAC
in a firmware with the name `wlan/qca_cld/wlan_mac.bin`.
That firmware is, by default, symlinked to
`/mnt/vendor/persist/wlan_mac.bin`.
ZTE has somewhat changed the qcacld-3.0 logic to load
the MAC from another place. ZTE's factory puts a plain-text
MAC in `/mnt/vendor/persist/wifimac.dat`.
Add a script to write the MAC address found in the file
created by factory to the expected location of qcacld-3.0.

This implementation is based on Xiaomi 8150
modified for the ZTE Axon 10 Pro.